+++ s3torchconnectorclient/python/tst/integration/test_logging.py
@@ -183,6 +183,8 @@ def _start_subprocess(
     process = subprocess.Popen(
         [
             sys.executable,
+            "-W",
+            "ignore::FutureWarning",
             "-c",
             PYTHON_TEST_CODE,
             image_directory.s3_uri,